+++ doc/CHANGELOG.md
@@ -31,6 +31,7 @@ See doc/COPYRIGHT.rdoc for more details.
 
 * `#3050` Fix: Fix: [Work Package Tracking] Internal error with query from the time before work packages
 * `#3202` Fix: [Bug] Grouping work packages by responsible is broken
+* `#3055` [Work Package Tracking] Former deleted planning elements not handled in any way
 
 ## 3.0.0pre34